requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.89.0
anyio==4.13.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.2
coloredlogs==15.0.1
cryptography==46.0.6
distro==1.9.0
docstring_parser==0.17.0
fastapi==0.135.3
fastembed==0.8.0
filelock==3.25.2
flatbuffers==25.12.19
fsspec==2026.3.0
greenlet==3.3.2
h11==0.16.0
hf-xet==1.4.3
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.9.0
humanfriendly==10.0
idna==3.11
Jinja2==3.1.6
jiter==0.13.0
joblib==1.5.3
loguru==0.7.3
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
mmh3==5.2.1
mpmath==1.3.0
networkx==3.6.1
numpy==1.26.4
onnxruntime==1.23.2
openai==2.30.0
packaging==26.0
pgvector==0.4.2
pillow==12.2.0
protobuf==7.34.1
psycopg2-binary==2.9.11
py_rust_stemmers==0.1.5
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
PyGithub==2.9.0
Pygments==2.20.0
PyJWT==2.12.1
PyNaCl==1.6.2
python-dotenv==1.2.2
PyYAML==6.0.3
regex==2026.4.4
requests==2.33.1
rich==14.3.3
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.1
shellingham==1.5.4
sniffio==1.3.1
SQLAlchemy==2.0.49
starlette==1.0.0
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.22.2
torch==2.2.2
tqdm==4.67.3
transformers==5.5.0
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.44.0
